Refactoring to put addXXXInfo in traits
